requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
playwright>=1.45.0
playwright-stealth>=1.0.6
motor>=3.6.0
pymongo>=4.9.0
groq>=0.11.0
pydantic>=2.8.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.31.0
readability-lxml>=0.8.1

# Phase 7 — Google Sheets CRM (optional, install if using sheets sync)
gspread>=6.0.0
google-auth>=2.28.0
